// src/users/user.controller.interface.ts
import { NextFunction, Request, Response } from 'express';

export interface UserControllerInterface {
	login: (req: Request, res: Response, next: NextFunction) => void;
	register: (req: Request, res: Response, next: NextFunction) => void;
	info: (req: Request, res: Response, next: NextFunction) => void;
}